Move CloudWatch Logs client to aws/client package
Moved internal/providers/aws/orchestrator/client.go to
internal/providers/aws/client/cloudwatchlogs.go for better
organization and consistency.

Rationale:
- The CloudWatch Logs client follows the same pattern as ECS and IAM
  clients (interface + adapter wrapper)
- The aws/client package is documented as the place for "AWS-specific
  client interfaces for AWS services"
- Separates infrastructure concerns from orchestrator business logic
- Improves discoverability and reusability

Changes:
- Created internal/providers/aws/client/cloudwatchlogs.go with
  CloudWatchLogsClient interface and adapter
- Updated imports in orchestrator package files to use
  awsClient.CloudWatchLogsClient
- Removed internal/providers/aws/orchestrator/client.go
