+++ internal/providers/aws/client/cloudwatchlogs.go
@@ -1,4 +1,4 @@
-package orchestrator
+package client
 
 import (
 	"context"
